+++ client/router.js
@@ -5,7 +5,7 @@ Vue.use(Router)
 
 const Home = () => import('~/pages/home').then(m => m.default || m)
 const Login = () => import('~/pages/login').then(m => m.default || m)
-// const Register = () => import('~/pages/auth/register').then(m => m.default || m)
+const Register = () => import('~/pages/register').then(m => m.default || m)
 // const PasswordReset = () => import('~/pages/auth/password/reset').then(m => m.default || m)
 // const PasswordRequest = () => import('~/pages/auth/password/email').then(m => m.default || m)
 
@@ -15,8 +15,8 @@ const Login = () => import('~/pages/login').then(m => m.default || m)
 
 const routes = [
   { path: '/', name: 'home', component: Home },
-  { path: '/login', name: 'login', component: Login }
-  // { path: '/register', name: 'register', component: Register },
+  { path: '/login', name: 'login', component: Login },
+  { path: '/register', name: 'register', component: Register }
   // { path: '/password/reset', name: 'password.request', component: PasswordRequest },
   // { path: '/password/reset/:token', name: 'password.reset', component: PasswordReset },